test pass for contract param serialize and deserialize

+++ source/ledger/ledger-model/src/main/java/com/jd/blockchain/ledger/CONTRACT_INT16.java
@@ -14,5 +14,5 @@ import com.jd.blockchain.consts.DataCodes;
 public interface CONTRACT_INT16 {
 
     @DataField(order=2, primitiveType= PrimitiveType.INT16)
-    int getValue();
+    short getValue();
 }
