+++ source/ledger/ledger-model/src/main/java/com/jd/blockchain/ledger/CONTRACT_INT8.java
@@ -14,5 +14,5 @@ import com.jd.blockchain.consts.DataCodes;
 public interface CONTRACT_INT8 {
 
     @DataField(order=2, primitiveType= PrimitiveType.INT8)
-    int getValue();
+    Byte getValue();
 }
